NamedParameterJdbc fix

+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/JdbcUserFilmLikeStorage.java
@@ -1,37 +1,36 @@
 package ru.yandex.practicum.filmorate.storage.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.storage.UserFilmLikeStorage;
-
-import java.time.LocalDate;
 
 @Repository
 public class JdbcUserFilmLikeStorage implements UserFilmLikeStorage {
 
-	private final JdbcTemplate jdbcTemplate;
+	private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
 	@Autowired
-	public JdbcUserFilmLikeStorage(JdbcTemplate jdbcTemplate) {
-		this.jdbcTemplate = jdbcTemplate;
+	public JdbcUserFilmLikeStorage(NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+		this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
 	}
 
 	@Override
 	public void addLike(Long filmId, Long userId) {
-		jdbcTemplate.update("INSERT INTO user_film_like (film_id, user_id, created_at) " +
-						"VALUES (?, ?, ?)",
-				filmId, userId, LocalDate.now());
+		String sql = "INSERT INTO user_film_like (film_id, user_id) VALUES (:filmId, :userId)";
+		MapSqlParameterSource params = new MapSqlParameterSource();
+		params.addValue("filmId", filmId);
+		params.addValue("userId", userId);
+		namedParameterJdbcTemplate.update(sql, params);
 	}
 
 	@Override
 	public void deleteLike(Long filmId, Long userId) {
-		jdbcTemplate.update(
-				"DELETE FROM user_film_like " +
-						"WHERE film_id = ? " +
-						"AND user_id = ?",
-				filmId,
-				userId
-		);
+		String sql = "DELETE FROM user_film_like WHERE film_id = :filmId AND user_id = :userId";
+		MapSqlParameterSource params = new MapSqlParameterSource();
+		params.addValue("filmId", filmId);
+		params.addValue("userId", userId);
+		namedParameterJdbcTemplate.update(sql, params);
 	}
 }
